refactor(gui): stage 5 — extract view/grid/tray and keyboard/menu modules
- src/app/view.js: showSingle, switchTo/switchToProject, updateAppTitle,
  renderGrid + layout state, spatial moves, minimize/restore + tray,
  renderEmptyState, setView (250ms snap intact), and the terms-host
  ResizeObserver — one module because tray/empty-state/view call each
  other (splitting would force injected cycles). ensureTerm, setActive,
  focusActiveTerm and the scroll tracer are injected via initView(deps).
- src/app/keyboard.js: the window keydown handler, menu action table +
  registrations, toggle/nav helpers, delete-project confirm path, and
  moveActiveSession. Font helpers + focusActiveTerm injected via
  initKeyboard(deps); everything else imported from sibling modules.

main.js: 2004 → 1249 lines. Suites: vitest 149, mock e2e 52(+1 skip),
e2e-real 5.
