+++ cmd/hivegui/frontend/src/main.js
@@ -52,7 +52,17 @@ import { openCommandPalette, initCommandPalette } from './app/modals/command-pal
 import { openHelpOverlay, closeHelpOverlay, toggleHelpOverlay, initHelpOverlay } from './app/modals/help-overlay.js';
 import { renderSidebar, updateSidebarSelection, initSidebar } from './app/sidebar.js';
 import { onSessionBell, clearAttention, wireDaemonEvents } from './app/events.js';
-import { isDaemonRestarting, manualUpdateCheck, initBanners } from './app/banners.js';
+import { isDaemonRestarting, initBanners } from './app/banners.js';
+import {
+  showSingle, switchTo, switchToProject, updateAppTitle, renderGrid,
+  gridSpatialMove, shiftActiveProject, gridScopeSessions, minimizeSession,
+  restoreSession, renderMinimizedTray, renderEmptyState, setView, initView,
+} from './app/view.js';
+import {
+  initKeyboard, toggleSidebar, toggleProjectGrid, toggleAllGrid,
+  confirmAndDeleteProject, deleteActiveProject, navSession, reorderActive,
+  switchToNthSession,
+} from './app/keyboard.js';
 
 // Scroll/replay tracer — gated on localStorage hive.debug = '1' (the
 // focus consistency checker's switch). The gate is latched here at
@@ -878,167 +888,6 @@ function ensureTerm(info) {
   return st;
 }
 
-function showSingle(id) {
-  termsHost.classList.add('single');
-  termsHost.classList.remove('grid');
-  // Hide everything except the active tile.
-  for (const [sid, st] of state.terms) {
-    if (sid === id) st.show();
-    else st.hide();
-    st.host.classList.remove('in-grid', 'active');
-  }
-  const st = id ? state.terms.get(id) : null;
-  if (st) st.ensureAttached();
-}
-
-function switchTo(id) {
-  if (id === state.activeId && state.view === 'single') {
-    focusActiveTerm();
-    return;
-  }
-  setActive(id);
-  let info = null;
-  if (id) {
-    info = state.sessions.find((s) => s.id === id);
-    if (info) ensureTerm(info);
-  }
-  // Retarget the grid scope if the new session belongs to a different
-  // project than the one currently shown in grid-project mode.
-  if (state.view === 'grid-project' && info) {
-    const pid = info.projectId ?? info.project_id;
-    if (pid && pid !== state.gridProjectId) state.gridProjectId = pid;
-  }
-  if (state.view === 'single') showSingle(id);
-  else renderGrid();
-  updateSidebarSelection();
-  setStatus(info ? info.name : '');
-  updateAppTitle();
-  // setActive() called focusActiveTerm() before ensureTerm() existed
-  // for a brand-new session — re-focus now that the SessionTerm is
-  // created and visible. Without this, typing after creating a
-  // session lands in whichever terminal had focus before.
-  if (id) focusActiveTerm();
-}
-
-// updateAppTitle composes "Hive — <session> — <termTitle>" and pushes
-// it to both document.title and the native window title bar. The
-// termTitle slot is whatever the running TUI most recently set via
-// the OSC 0/2 escape sequence; empty if the program never set one.
-//
-// Throttled with a trailing-edge timer: programs like fish prompts
-// or progress encoders can fire OSC 0/2 dozens of times per second,
-// and each WindowSetTitle is a Wails IPC round-trip. 100ms keeps the
-// title visibly responsive without flooding the bridge.
-let _appTitleTimer = null;
-function updateAppTitle() {
-  if (_appTitleTimer) return;
-  _appTitleTimer = setTimeout(() => {
-    _appTitleTimer = null;
-    const id = state.activeId;
-    const info = id ? state.sessions.find((s) => s.id === id) : null;
-    const parts = ['Hive'];
-    if (info?.name) parts.push(info.name);
-    const t = id ? state.terms.get(id) : null;
-    if (t?.termTitle && t.termTitle !== info?.name) parts.push(t.termTitle);
-    const title = parts.join(' — ');
-    document.title = title;
-    try { WindowSetTitle(title); } catch (_) { /* runtime not ready */ }
-  }, 100);
-}
-
-// switchToProject activates a project: in grid-project mode it
-// retargets the grid, and in any mode it makes the project's first
-// session the active one. Empty projects are still selectable —
-// currentProjectId is set so ⌘N targets them correctly.
-function switchToProject(pid) {
-  if (!pid) return;
-  state.currentProjectId = pid;
-  if (state.view === 'grid-project') state.gridProjectId = pid;
-  const sessions = state.sessions
-    .filter((s) => (s.projectId ?? s.project_id) === pid)
-    .sort((a, b) => (a.order ?? 0) - (b.order ?? 0));
-  if (sessions[0]) switchTo(sessions[0].id);
-  else {
-    state.activeId = null;
-    if (state.view === 'single') showSingle(null);
-    else renderGrid();
-    updateSidebarSelection();
-  }
-}
-
-// gridLayout caches the (rows, cols) chosen for the current scope plus
-// the per-tile placement so the keyboard navigation logic doesn't have
-// to recompute. assignments[i] = { row, col, rowSpan } — tiles above
-// last-row empty cells extend downward to fill the grid (matches
-// current Hive's behavior). cellMap[row*cols + col] = session index.
-let gridLayout = { rows: 1, cols: 1, sessions: [], assignments: [], cellMap: [] };
-
-// renderGrid lays out every tile that should be visible in the
-// current grid scope. Tiles for other sessions are hidden but kept
-// alive (so their xterm scrollback persists across mode switches).
-function renderGrid() {
-  termsHost.classList.remove('single');
-  termsHost.classList.add('grid');
-  const gridSessions = gridScopeSessions();
-  const gridIDs = new Set(gridSessions.map((s) => s.id));
-  const n = gridSessions.length;
-
-  // Ensure every grid session has a SessionTerm and is attached.
-  // Move tiles into the desired DOM order (row-major) so that flexbox
-  // / CSS grid honors the navigation order without us having to set
-  // grid-row/column explicitly.
-  for (const info of gridSessions) {
-    const st = ensureTerm(info);
-    st.host.classList.add('in-grid');
-    st.host.classList.toggle('active', info.id === state.activeId);
-    st.host.classList.toggle('attention', state.attention.has(info.id));
-    st.ensureAttached();
-    termsHost.appendChild(st.host); // re-order to keep DOM == nav order
-  }
-  // Hide / unmark tiles outside the scope.
-  for (const [sid, st] of state.terms) {
-    if (!gridIDs.has(sid)) {
-      st.host.classList.remove('in-grid', 'active');
-      st.host.style.gridRow = '';
-      st.host.style.gridColumn = '';
-    }
-  }
-
-  // Pick (rows, cols) that fills the container, then derive
-  // per-tile assignments + the cellMap used by spatial nav. Pure
-  // logic lives in lib/grid.js so it can be unit-tested.
-  const w = termsHost.clientWidth || 800;
-  const h = termsHost.clientHeight || 600;
-  const { rows, cols, assignments, cellMap } = buildGridLayout(n, w, h);
-
-  termsHost.style.gridTemplateColumns = `repeat(${cols}, 1fr)`;
-  termsHost.style.gridTemplateRows = `repeat(${rows}, 1fr)`;
-
-  // Apply each tile's row span. CSS grid 1-based; row indices are
-  // implicit row-major, so we only need to span when rowSpan > 1.
-  for (let i = 0; i < n; i++) {
-    const a = assignments[i];
-    const st = state.terms.get(gridSessions[i].id);
-    if (!st) continue;
-    if (a.rowSpan > 1) {
-      st.host.style.gridRow = `span ${a.rowSpan}`;
-    } else {
-      st.host.style.gridRow = '';
-    }
-    st.host.style.gridColumn = '';
-  }
-
-  gridLayout = { rows, cols, sessions: gridSessions, assignments, cellMap };
-
-  // No explicit refit pass: each tile's ResizeObserver fires when its
-  // body box changes (CSS grid cell resized, in-grid class toggled,
-  // tile shown/hidden). That's the only place fit.fit() runs.
-}
-
-// setActive centralizes "the focused session changed" so every code
-// path (click, arrow nav, project switch, switchTo) clears the bell
-// indicator the same way and syncs the current project to whatever
-// project the new session belongs to.
 function setActive(id) {
   if (id) {
     state.attention.delete(id);
@@ -1259,567 +1108,6 @@ function refocusActiveTerm() {
 // Uses cellMap to honor row-spanned tiles: e.g. with 3 sessions in a
 // 2x2 grid the bottom-right cell is absorbed by tile 1, so pressing
 // "right" from tile 2 lands on tile 1 instead of doing nothing.
-function gridSpatialMove(dCol, dRow) {
-  const { sessions } = gridLayout;
-  if (sessions.length === 0) return;
-  const idx = sessions.findIndex((s) => s.id === state.activeId);
-  if (idx < 0) {
-    setActive(sessions[0].id);
-    renderGrid();
-    updateSidebarSelection();
-    return;
-  }
-  const target = computeSpatialMove(gridLayout, idx, dCol, dRow);
-  if (target == null) return;
-  setActive(sessions[target].id);
-  renderGrid();
-  updateSidebarSelection();
-  setStatus(sessions[target].name);
-}
-
-function shiftActiveProject(delta) {
-  if (state.projects.length === 0) return;
-  const cur = activeProjectId();
-  const i = state.projects.findIndex((p) => p.id === cur);
-  if (i < 0) return;
-  const next = state.projects[(i + delta + state.projects.length) % state.projects.length];
-  state.currentProjectId = next.id;
-  if (state.view === 'grid-project') state.gridProjectId = next.id;
-
-  const sessions = state.sessions
-    .filter((s) => (s.projectId ?? s.project_id) === next.id)
-    .sort((a, b) => (a.order ?? 0) - (b.order ?? 0));
-  if (sessions[0]) {
-    setActive(sessions[0].id);
-  } else {
-    // Empty project — keep the project selected but drop the active
-    // session so the user can ⌘N into it. activeProjectId() now
-    // returns the empty project because currentProjectId is set.
-    state.activeId = null;
-  }
-  if (state.view === 'single') showSingle(state.activeId);
-  else renderGrid();
-  updateSidebarSelection();
-  setStatus(`${next.name}${sessions.length === 0 ? ' (empty)' : ''}`);
-}
-
-// gridScopeSessions returns the list of sessions that should be tiled
-// in the current grid view.
-function gridScopeSessions() {
-  if (state.view === 'grid-all') {
-    return filterMinimized(orderedSessions(), state.minimized);
-  }
-  if (state.view === 'grid-project') {
-    const pid = state.gridProjectId || activeProjectId();
-    const scoped = state.sessions
-      .filter((s) => (s.projectId ?? s.project_id) === pid)
-      .sort((a, b) => (a.order ?? 0) - (b.order ?? 0));
-    return filterMinimized(scoped, state.minimized);
-  }
-  return [];
-}
-
-// minimizeSession hides a session from grid views by adding its id to
-// state.minimized. The session stays alive; its tile is removed on the
-// next renderGrid(). Single-session mode is unaffected — the user can
-// still switch to a minimized session via the sidebar / palette / ⌘[/].
-function minimizeSession(id) {
-  if (!id || state.minimized.has(id)) return;
-  state.minimized.add(id);
-  // If the active session is the one being minimized while in grid
-  // mode, hand focus to the next still-visible session so the focus
-  // ring doesn't vanish onto an offscreen tile.
-  if (state.activeId === id && state.view !== 'single') {
-    const next = gridScopeSessions().find((s) => s.id !== id);
-    if (next) setActive(next.id);
-  }
-  if (state.view !== 'single') {
-    renderGrid();
-    rebaselineGridReplayCols();
-  }
-  renderMinimizedTray();
-}
-
-// restoreSession removes a session from state.minimized and switches
-// to it. Works from any view — switchTo handles the view-aware repaint.
-function restoreSession(id) {
-  if (!id) return;
-  state.minimized.delete(id);
-  renderMinimizedTray();
-  switchTo(id);
-  if (state.view !== 'single') {
-    rebaselineGridReplayCols();
-  }
-}
-
-// rebaselineGridReplayCols defers a baseline reset to after the next
-// two animation frames. The first rAF lets the CSS grid layout settle
-// and ResizeObserver fire _onBodyResize on each affected tile (which
-// updates this.term.cols via fit.fit() and may arm a 100ms replay
-// debounce). The second rAF then snapshots the new term.cols as the
-// baseline and clears the pending debounce — turning a layout-driven
-// width change into a no-op rather than a spurious scrollback replay.
-// Pure user window resizes still flow through the threshold path in
-// _onBodyResize and continue to request replays as before.
-function rebaselineGridReplayCols() {
-  requestAnimationFrame(() => requestAnimationFrame(() => {
-    for (const st of state.terms.values()) {
-      if (st.host.classList.contains('in-grid')) {
-        st.rebaselineReplayCols('layout');
-      }
-    }
-  }));
-}
-
-// renderMinimizedTray rebuilds the #minimized-tray chip row from
-// state.minimized. Hidden when the set is empty.
-function renderMinimizedTray() {
-  const tray = document.getElementById('minimized-tray');
-  if (!tray) return;
-  tray.innerHTML = '';
-  const ids = Array.from(state.minimized);
-  if (ids.length === 0) {
-    tray.classList.add('hidden');
-    return;
-  }
-  tray.classList.remove('hidden');
-  // Preserve display order so the chip row reads top-to-bottom like
-  // the sidebar / grid would.
-  const ord = orderedSessions().filter((s) => state.minimized.has(s.id));
-  for (const info of ord) {
-    const chip = document.createElement('button');
-    chip.className = 'min-chip';
-    chip.type = 'button';
-    chip.dataset.sid = info.id;
-    chip.title = `Restore ${info.name}`;
-    chip.setAttribute('aria-label', `Restore ${info.name}`);
-    chip.style.setProperty('--session-color', info.color || '#888');
-
-    const dot = document.createElement('span');
-    dot.className = 'min-chip-color';
-    chip.append(dot);
-
-    const name = document.createElement('span');
-    name.className = 'min-chip-name';
-    name.textContent = info.name;
-    chip.append(name);
-
-    const pid = info.projectId ?? info.project_id;
-    const proj = state.projects.find((p) => p.id === pid);
-    if (proj?.name) {
-      const projLabel = document.createElement('span');
-      projLabel.className = 'min-chip-project';
-      projLabel.textContent = proj.name;
-      chip.append(projLabel);
-    }
-
-    chip.addEventListener('click', (e) => {
-      e.stopPropagation();
-      restoreSession(info.id);
-    });
-    tray.append(chip);
-  }
-  // Minimize/restore changes which sessions are visible without a
-  // sidebar render — re-evaluate the empty state here too.
-  renderEmptyState();
-}
-
-// renderEmptyState shows an actionable hint pane when the current
-// scope has nothing to display (first run, empty project, everything
-// minimized). Pure model in lib/empty-state.js; this just projects it
-// onto the #empty-state element. Cheap enough to call from every
-// repaint path — DOM is rebuilt only when the model changes.
-function renderEmptyState() {
-  const el = document.getElementById('empty-state');
-  if (!el) return;
-  const model = emptyStateModel({
-    projects: state.projects,
-    sessions: state.sessions,
-    view: state.view,
-    currentProjectId: state.currentProjectId,
-    gridProjectId: state.gridProjectId,
-    minimized: state.minimized,
-    isMac,
-  });
-  if (!model) {
-    el.classList.add('hidden');
-    el.dataset.kind = '';
-    delete el.dataset.sig;
-    return;
-  }
-  // Key the rebuild off the full model, not just the kind: within
-  // 'first-run' the hint/actions vary with projects.length, so a
-  // kind-only check would leave stale text and buttons behind.
-  const sig = JSON.stringify(model);
-  if (el.dataset.sig !== sig) {
-    el.dataset.sig = sig;
-    el.dataset.kind = model.kind;
-    el.innerHTML = '';
-    const title = document.createElement('div');
-    title.className = 'empty-title';
-    title.textContent = model.title;
-    const hint = document.createElement('div');
-    hint.className = 'empty-hint';
-    hint.textContent = model.hint;
-    el.append(title, hint);
-    if (model.actions.length) {
-      const row = document.createElement('div');
-      row.className = 'empty-actions';
-      for (const a of model.actions) {
-        const btn = document.createElement('button');
-        btn.type = 'button';
-        btn.textContent = a.label;
-        btn.addEventListener('click', (e) => {
-          // The launcher now opens synchronously; without this, the
-          // same click bubbles to the document-level outside-click
-          // closer and shuts it in the same tick.
-          e.stopPropagation();
-          if (a.id === 'new-session') openLauncher();
-          else if (a.id === 'new-project') openProjectEditor(null);
-        });
-        row.appendChild(btn);
-      }
-      el.appendChild(row);
-    }
-  }
-  el.classList.remove('hidden');
-}
-
-function setView(view) {
-  state.view = view;
-  try { localStorage.setItem(VIEW_STORAGE_KEY, view); } catch {}
-  if (view === 'grid-project') {
-    state.gridProjectId = activeProjectId();
-  }
-  if (view === 'single') {
-    showSingle(state.activeId);
-  } else {
-    renderGrid();
-  }
-  updateSidebarSelection();
-  // Toggling grid/fullscreen via the menu blurs the xterm; restore
-  // focus so the user can keep typing into the active session.
-  focusActiveTerm();
-  // Mode switches are deliberate user actions — always snap visible
-  // tiles to the bottom. Without this, xterm lands wherever the
-  // buffer happened to be (often mid-history), which is jarring.
-  //
-  // Defer past focusActiveTerm's full focus-retry budget
-  // (FOCUS_MAX_RETRIES = 8 frames ≈ 130ms) before snapping. xterm's
-  // scrollToBottom() refreshes the renderer, which can fire focusout
-  // on the helper textarea — synchronous snap broke focus on Linux,
-  // single-rAF snap broke focus on macOS, because each platform's
-  // rAF cadence races focusActiveTerm's retry loop differently.
-  // A 250ms setTimeout clears the polling window on every platform;
-  // a quarter-second pre-snap pause is below the perception threshold
-  // for visual settling after a mode change.
-  setTimeout(() => {
-    if (scrollTrace.rec.enabled) {
-      scrollTrace.rec('mode-snap', { view });
-    }
-    snapVisibleTermsToBottom(state.terms.values());
-  }, 250);
-  const ord = orderedSessions();
-  const active = ord.find((s) => s.id === state.activeId);
-  setStatus(`${view}${active ? ' • ' + active.name : ''}`);
-  renderEmptyState();
-}
-
-
-// ---------- keyboard ----------
-
-window.addEventListener('keydown', (e) => {
-  if (!launcherEl.classList.contains('hidden')) {
-    const handle = (fn) => { e.preventDefault(); e.stopPropagation(); fn(); };
-    if (e.key === 'ArrowDown' || (e.key === 'Tab' && !e.shiftKey)) return handle(() => moveLauncherSelection(+1));
-    if (e.key === 'ArrowUp'   || (e.key === 'Tab' && e.shiftKey))  return handle(() => moveLauncherSelection(-1));
-    if (e.key === 'Enter')   return handle(activateLauncherSelection);
-    if (e.key === 'Escape')  return handle(closeLauncher);
-    if (cmdOrCtrl(e) && (e.key === 'n' || e.key === 'N')) return handle(closeLauncher);
-    // Digit shortcut: 1–9 picks the corresponding row. Skipped when
-    // a modifier is held so things like ⌘1 (browser tab switch) and
-    // ⌘+ aren't swallowed.
-    if (!e.metaKey && !e.ctrlKey && !e.altKey && /^[1-9]$/.test(e.key)) {
-      const i = parseInt(e.key, 10) - 1;
-      if (i < launcherState.items.length) {
-        return handle(() => {
-          launcherState.selected = i;
-          activateLauncherSelection();
-        });
-      }
-    }
-  }
-  if (!editorEl.classList.contains('hidden')) {
-    return; // editor's own listener handles keys
-  }
-  const _palette = document.getElementById('command-palette');
-  if (_palette && !_palette.classList.contains('hidden')) {
-    return; // palette's own listener handles keys
-  }
-  const _help = document.getElementById('help-overlay');
-  if (_help && !_help.classList.contains('hidden')) {
-    if (e.key === 'Escape' || (cmdOrCtrl(e) && e.key === '/')) {
-      e.preventDefault();
-      e.stopPropagation();
-      closeHelpOverlay();
-    } else if (e.key === 'Tab') {
-      // aria-modal promises focus stays inside the dialog. The close
-      // button is its only focusable element, so trap Tab on it —
-      // otherwise focus walks into the page (eventually a hidden
-      // terminal's textarea) and keystrokes leak behind the backdrop.
-      e.preventDefault();
-      e.stopPropagation();
-      document.getElementById('help-overlay-close')?.focus();
-    }
-    return; // overlay owns the keyboard while open
-  }
-
-  // Dead-session overlay: route Enter/Escape to the active session's
-  // overlay if it's shown. In grid mode the user can still click any
-  // tile's buttons directly; this just handles the focused tile.
-  if (state.activeId) {
-    const t = state.terms.get(state.activeId);
-    if (t?.deadOverlayShown) {
-      if (e.key === 'Enter') { e.preventDefault(); e.stopPropagation(); t._closeDead(); return; }
-      if (e.key === 'Escape') { e.preventDefault(); e.stopPropagation(); t._dismissDead(); return; }
-    }
-  }
-
-  const swallow = () => { e.preventDefault(); e.stopPropagation(); };
-
-  // Ctrl+` opens an OS terminal at the active session's worktree.
-  // Mirrors VS Code; intentionally Ctrl on every platform — macOS
-  // reserves ⌘` for native window cycling, so we never bind to it.
-  // Handled before the ⌘/Ctrl gate below so it fires on mac too.
-  if (e.ctrlKey && !e.metaKey && !e.altKey && !e.shiftKey && e.code === 'Backquote') {
-    swallow();
-    OpenTerminalAt(activeCwd()).catch(reportFailure('open terminal'));
-    return;
-  }
-
-  const meta = cmdOrCtrl(e);
-  if (!meta) return;
-
-  if (e.key === '=' || e.key === '+') {
-    swallow();
-    bumpFontSize(+1);
-    return;
-  }
-  if (e.key === '-' || e.key === '_') {
-    swallow();
-    bumpFontSize(-1);
-    return;
-  }
-  if (e.key === '0') {
-    swallow();
-    resetFontSize();
-    return;
-  }
-
-  if ((e.key === 'k' || e.key === 'K') && e.shiftKey) {
-    swallow();
-    openCommandPalette();
-    return;
-  }
-  if (e.key === '/') {
-    swallow();
-    openHelpOverlay();
-    return;
-  }
-  if (e.key === 'p' || e.key === 'P') {
-    swallow();
-    if (e.shiftKey) duplicateActiveSessionChooseTool();
-    else duplicateActiveSession();
-  } else if (e.key === 't' || e.key === 'T') {
-    swallow();
-    if (e.shiftKey) openLauncher(undefined, { forceWorktree: true });
-    else openLauncher();
-  } else if (e.key === 'Backspace' && e.shiftKey) {
-    swallow();
-    deleteActiveProject();
-  } else if (e.key === 's' || e.key === 'S') {
-    swallow();
-    // Route through toggleSidebar so the keyboard path stays in
-    // lockstep with the menu / command-palette path (including the
-    // post-reflow refocus added for #208 R3). Inline class flips
-    // here previously skipped the refocus and stranded keystrokes
-    // on document.body after a prior window resize.
-    toggleSidebar();
-  } else if (e.key === 'g' || e.key === 'G') {
-    swallow();
-    if (e.shiftKey) {
-      setView(state.view === 'grid-all' ? 'single' : 'grid-all');
-    } else {
-      setView(state.view === 'grid-project' ? 'single' : 'grid-project');
-    }
-  } else if (e.key === 'Enter') {
-    // ⌘Enter mirrors ⌘G: in a grid mode it maximizes the active
-    // tile back to single mode; in single mode it expands to a
-    // per-project grid for context.
-    swallow();
-    if (state.view === 'single') setView('grid-project');
-    else setView('single');
-  } else if (e.key === 'n' || e.key === 'N') {
-    swallow();
-    if (e.shiftKey) {
-      OpenNewWindow().catch(reportFailure('new window'));
-    } else {
-      // ⌘N — new project. (⌥⌘N is reserved by macOS Spotlight.)
-      openProjectEditor(null);
-    }
-  } else if (e.key === 'w' || e.key === 'W') {
-    swallow();
-    if (e.shiftKey) {
-      CloseWindow().catch(reportFailure('close window'));
-    } else if (state.activeId) {
-      // force=false: lets the daemon refuse with worktree_dirty if
-      // the worktree has uncommitted changes; the control:error
-      // handler then shows a confirm dialog and retries with force.
-      KillSession(state.activeId, false).catch(reportFailure('close'));
-    }
-  } else if (/^[1-9]$/.test(e.key)) {
-    const idx = parseInt(e.key, 10) - 1;
-    const ord = orderedSessions();
-    if (idx < ord.length) {
-      swallow();
-      switchTo(ord[idx].id);
-    }
-  } else if (e.key === 'ArrowLeft') {
-    swallow();
-    if (state.view !== 'single') gridSpatialMove(-1, 0);
-    else moveActiveSession(-1, e.shiftKey);
-  } else if (e.key === 'ArrowRight') {
-    swallow();
-    if (state.view !== 'single') gridSpatialMove(+1, 0);
-    else moveActiveSession(+1, e.shiftKey);
-  } else if (e.key === 'ArrowUp') {
-    swallow();
-    if (state.view !== 'single') gridSpatialMove(0, -1);
-    else moveActiveSession(-1, e.shiftKey);
-  } else if (e.key === 'ArrowDown') {
-    swallow();
-    if (state.view !== 'single') gridSpatialMove(0, +1);
-    else moveActiveSession(+1, e.shiftKey);
-  } else if (e.key === '[') {
-    swallow();
-    shiftActiveProject(-1);
-  } else if (e.key === ']') {
-    swallow();
-    shiftActiveProject(+1);
-  }
-}, true);
-
-// ---------- menu actions ----------
-//
-// Native menu items emit `menu:<action>` events from cmd/hivegui/menu.go.
-// They dispatch to the same handlers as the keyboard listener above so the
-// menu and keyboard stay in lockstep — when you add a shortcut, add it
-// here AND in menu.go.
-
-function toggleSidebar() {
-  document.getElementById('app').classList.toggle('sidebar-hidden');
-  // Layout reflow → tile bodies resize → ResizeObserver fits xterm,
-  // and fit.fit() can synchronously fire focusout on the helper-
-  // textarea as the canvas re-sizes. Without re-asserting focus,
-  // keystrokes strand on document.body even though the visual
-  // .term-focused stays correctly pinned on the active tile (#208 R3).
-  //
-  // Sync-fire focusActiveTerm so setFocusedTile's rAF retry loop is
-  // armed before the first focusout; staggered delayed re-fires catch
-  // any focusout that escapes the standard 8-frame retry budget
-  // (later RO callbacks, WebGL canvas swap, DPR settle). All calls
-  // are idempotent — re-focusing an already-focused element is a no-op.
-  focusActiveTerm();
-  setTimeout(() => focusActiveTerm(), 32);
-  setTimeout(() => focusActiveTerm(), 100);
-  setTimeout(() => focusActiveTerm(), 250);
-}
-
-function toggleProjectGrid() {
-  setView(state.view === 'grid-project' ? 'single' : 'grid-project');
-}
-
-function toggleAllGrid() {
-  setView(state.view === 'grid-all' ? 'single' : 'grid-all');
-}
-
-function navSession(delta) {
-  if (state.view !== 'single') {
-    gridSpatialMove(delta > 0 ? +1 : -1, 0);
-  } else {
-    moveActiveSession(delta, false);
-  }
-}
-
-function reorderActive(delta) {
-  if (state.view === 'single') moveActiveSession(delta, true);
-  else gridSpatialMove(delta > 0 ? +1 : -1, 0);
-}
-
-function switchToNthSession(n) {
-  const ord = orderedSessions();
-  if (n - 1 < ord.length) switchTo(ord[n - 1].id);
-}
-
-const menuActions = {
-  'menu:new-session': () => openLauncher(),
-  'menu:new-session-worktree': () => openLauncher(undefined, { forceWorktree: true }),
-  'menu:duplicate-session': duplicateActiveSession,
-  'menu:duplicate-session-choose-tool': duplicateActiveSessionChooseTool,
-  'menu:restart-session': restartActiveSession,
-  'menu:new-project': () => openProjectEditor(null),
-  'menu:delete-project': () => deleteActiveProject(),
-  'menu:command-palette': () => openCommandPalette(),
-  'menu:close-session': () => { if (state.activeId) KillSession(state.activeId, false).catch(reportFailure('close')); },
-  'menu:zoom-in': () => bumpFontSize(+1),
-  'menu:zoom-out': () => bumpFontSize(-1),
-  'menu:zoom-reset': () => resetFontSize(),
-  'menu:toggle-sidebar': toggleSidebar,
-  'menu:toggle-project-grid': toggleProjectGrid,
-  'menu:toggle-all-grid': toggleAllGrid,
-  'menu:next-session': () => navSession(+1),
-  'menu:prev-session': () => navSession(-1),
-  'menu:move-session-forward': () => reorderActive(+1),
-  'menu:move-session-backward': () => reorderActive(-1),
-  'menu:next-project': () => shiftActiveProject(+1),
-  'menu:prev-project': () => shiftActiveProject(-1),
-  'menu:check-for-updates': () => manualUpdateCheck(),
-  // Must toggle, not just open: the native ⌘/ accelerator intercepts
-  // the key before the webview on macOS, so the keydown close path
-  // (Escape/⌘/ in the window listener) never sees ⌘/ while the menu
-  // owns it.
-  'menu:keyboard-shortcuts': () => toggleHelpOverlay(),
-};
-for (const [name, fn] of Object.entries(menuActions)) {
-  EventsOn(name, fn);
-}
-for (let i = 1; i <= 9; i++) {
-  EventsOn(`menu:switch-${i}`, () => switchToNthSession(i));
-}
-
-// ---------- delete project ----------
-
-// confirmAndDeleteProject is the single confirm + KillProject path
-// shared by the sidebar ✕ button and the ⇧⌘⌫ shortcut. Kept as one
-// function so the prompt text and killSessions logic can't drift.
-async function confirmAndDeleteProject(proj) {
-  if (!proj) return;
-  const sessions = state.sessions.filter(
-    (s) => (s.projectId ?? s.project_id) === proj.id,
-  );
-  const msg = sessions.length
-    ? `Delete project "${proj.name}" and kill ${sessions.length} session${sessions.length === 1 ? '' : 's'}?`
-    : `Delete project "${proj.name}"?`;
-  const ok = await Confirm('Delete project', msg);
-  if (!ok) return;
-  KillProject(proj.id, sessions.length > 0).catch(reportFailure('delete project'));
-}
-
-function deleteActiveProject() {
-  const pid = activeProjectId();
-  confirmAndDeleteProject(state.projects.find((p) => p.id === pid));
-}
-
 // ---------- command palette ----------
 
 // Shortcut strings come from lib/shortcuts.js so the palette and the
@@ -1867,6 +1155,8 @@ initCommandPalette({ commands: paletteCommands, focusActiveTerm });
 initHelpOverlay({ setFocusedTile, focusActiveTerm });
 initSidebar({ switchTo, switchToProject, confirmAndDeleteProject, renderEmptyState, refocusActiveTerm });
 initBanners();
+initView({ ensureTerm, setActive, focusActiveTerm, scrollTrace });
+initKeyboard({ bumpFontSize, resetFontSize, focusActiveTerm });
 wireDaemonEvents({
   switchTo, renderMinimizedTray, updateAppTitle, focusActiveTerm,
   refocusActiveTerm, isDaemonRestarting, scrollTrace,
@@ -1880,51 +1170,6 @@ if (footerHintsEl) footerHintsEl.textContent = footerHints({ isMac });
 
 // moveActiveSession walks the (project_order, session_order) list.
 // reorder=true moves the session within its project only.
-function moveActiveSession(delta, reorder) {
-  const ord = orderedSessions();
-  const n = ord.length;
-  if (n === 0) return;
-  const idx = ord.findIndex((s) => s.id === state.activeId);
-  if (idx < 0) {
-    switchTo(ord[0].id);
-    return;
-  }
-  if (reorder) {
-    const cur = state.sessions.find((s) => s.id === state.activeId);
-    const sib = state.sessions
-      .filter((s) => (s.projectId ?? s.project_id) === (cur.projectId ?? cur.project_id))
-      .sort((a, b) => (a.order ?? 0) - (b.order ?? 0));
-    const sIdx = sib.findIndex((s) => s.id === state.activeId);
-    const next = (sIdx + delta + sib.length) % sib.length;
-    UpdateSession(state.activeId, '', '', next).catch(reportFailure('reorder'));
-    return;
-  }
-  const next = (idx + delta + n) % n;
-  switchTo(ord[next].id);
-}
-
-// ---------- resize ----------
-//
-// Per-tile fit is driven by each SessionTerm's own ResizeObserver
-// on its body. The only thing left at the page level is re-picking
-// (rows, cols) for the grid when the *container* changes shape —
-// e.g. landscape ↔ portrait window or sidebar drag — so tiles flow
-// from "side-by-side" to "stacked" and back.
-//
-// rAF coalesces the burst of RO entries during a continuous drag
-// into one renderGrid per frame. The guard also dodges the dreaded
-// "ResizeObserver loop completed with undelivered notifications"
-// warning that fires when a callback synchronously mutates layout.
-let _gridReflowQueued = false;
-new ResizeObserver(() => {
-  if (state.view === 'single' || _gridReflowQueued) return;
-  _gridReflowQueued = true;
-  requestAnimationFrame(() => {
-    _gridReflowQueued = false;
-    if (state.view !== 'single') renderGrid();
-  });
-}).observe(termsHost);
-
 // ---------- sidebar resize ----------
 //
 // Drag the right edge of the sidebar to resize. Width persists across
